[FEATURE] Add dual processor lib using shared memory for non-OS systems
This library is designed for system where the user layer and kernel
layer of openPOWERLINK run on different processors and shared memory
is used as the interface between the two processors.

The dual processor library requires the hardware to support two types
of memory: Common and dynamic memory.

The common memory can be used for exchanging status, errors, states
etc. and is used for creating dynamic memory mapping table used to
share the address for dynamically allocated shared memories.

Dynamic memory can be allocated at run time by any one of the processor
and these addresses are made available to the other processor through
the dynamic memory mapping table.

The library also provides support for handling interrupts for such
systems.