[TASK] Add flow control possiblity to SDO sequence layer
Flow control possibility for application by dropping
newly received sequences in the sequence layer and not
forwarding them to the command layer until the command
layer sends a "delayed" response.
The sequence layer will keep the connection alive
(resetting timeout of other node) by responding on an
acknowledge request with a retransmission request.
The own timeout will occur, if the own command layer
does not respond.

Change-Id: I31a0774b4679bbbb37b36b6403b316aeade1e6a7